chalice-api/requirements.txt
async-timeout==4.0.2
attrs==21.4.0
blessed==1.19.1
boto3==1.24.56
botocore==1.27.56
certifi==2022.6.15
cffi==1.15.1
chalice==1.27.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
decorator==5.1.1
Deprecated==1.2.13
idna==3.3
inquirer==2.10.0
jmespath==1.0.1
joblib==1.1.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.2
packaging==21.3
psutil==5.9.1
pycparser==2.21
pyemd==0.5.1
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
PyYAML==6.0
readchar==4.0.2
redis==4.3.4
s3transfer==0.6.0
scipy==1.9.0
six==1.16.0
tblib==1.7.0
tqdm==4.64.0
urllib3==1.26.11
wcwidth==0.2.5
wrapt==1.14.1
